+++ lambdas/data-migration2/tests/test-pdrs.js
@@ -22,19 +22,18 @@ const {
   recursivelyDeleteS3Bucket,
 } = require('@cumulus/aws-client/S3');
 const { dynamodbDocClient } = require('@cumulus/aws-client/services');
-const { RecordAlreadyMigrated } = require('@cumulus/errors');
+const { RecordAlreadyMigrated, PostgresUpdateFailed } = require('@cumulus/errors');
 
 // eslint-disable-next-line node/no-unpublished-require
 const { migrationDir } = require('../../db-migration');
 const { migratePdrRecord, migratePdrs } = require('../dist/lambda/pdrs');
 
-const generateTestPdr = (collection, provider, executionArn) => ({
+const buildCollectionId = (name, version) => `${name}___${version}`;
+
+const generateTestPdr = (params) => ({
   pdrName: cryptoRandomString({ length: 5 }),
-  collectionId: `${collection.name}___${collection.version}`,
-  provider: provider,
   status: 'running',
   progress: 10,
-  execution: executionArn,
   PANSent: false,
   PANmessage: 'message',
   stats: { total: 1, completed: 0, failed: 0, processing: 1 },
@@ -44,6 +43,7 @@ const generateTestPdr = (collection, provider, executionArn) => ({
   duration: 10,
   createdAt: Date.now(),
   updatedAt: Date.now(),
+  ...params,
 });
 
 let collectionsModel;
@@ -138,7 +138,11 @@ test.serial('migratePdrRecord correctly migrates PDR record', async (t) => {
   );
   const executionCumulusId = executionResponse[0];
 
-  const testPdr = generateTestPdr(testCollection, testProvider.name, execution.arn);
+  const testPdr = generateTestPdr({
+    collectionId: buildCollectionId(testCollection.name, testCollection.version),
+    provider: testProvider.name,
+    execution: execution.arn,
+  });
   await migratePdrRecord(testPdr, knex);
 
   const record = await pdrPgModel.get(knex, { name: testPdr.pdrName });
@@ -168,7 +172,10 @@ test.serial('migratePdrRecord correctly migrates PDR record', async (t) => {
 test.serial('migratePdrRecord throws SchemaValidationError on invalid source data from DynamoDB', async (t) => {
   const { knex, testCollection, testProvider } = t.context;
 
-  const testPdr = generateTestPdr(testCollection, testProvider.name);
+  const testPdr = generateTestPdr({
+    collectionId: buildCollectionId(testCollection.name, testCollection.version),
+    provider: testProvider.name,
+  });
 
   delete testPdr.status;
 
@@ -188,7 +195,10 @@ test.serial('migratePdrRecord handles nullable fields on source PDR data', async
     testProvider,
   } = t.context;
 
-  const testPdr = generateTestPdr(testCollection, testProvider.name);
+  const testPdr = generateTestPdr({
+    collectionId: buildCollectionId(testCollection.name, testCollection.version),
+    provider: testProvider.name,
+  });
 
   delete testPdr.execution;
   delete testPdr.PANSent;
@@ -225,16 +235,86 @@ test.serial('migratePdrRecord handles nullable fields on source PDR data', async
   );
 });
 
-test.serial('migratePdrRecord throws RecordAlreadyMigrated error for already migrated record', async (t) => {
+test.serial('migratePdrRecord throws RecordAlreadyMigrated error if previously migrated record is newer', async (t) => {
   const { knex, testCollection, testProvider } = t.context;
 
-  const testPdr = generateTestPdr(testCollection, testProvider.name);
+  const testPdr = generateTestPdr({
+    collectionId: buildCollectionId(testCollection.name, testCollection.version),
+    provider: testProvider.name,
+  });
   await migratePdrRecord(testPdr, knex);
 
+  const olderTestPdr = {
+    ...testPdr,
+    updatedAt: Date.now() - 1000,
+  };
+
   await t.throwsAsync(
-    migratePdrRecord(testPdr, knex),
+    migratePdrRecord(olderTestPdr, knex),
     { instanceOf: RecordAlreadyMigrated }
   );
+});
+
+test.serial('migratePdrRecord throws error if upsert does not return any rows', async (t) => {
+  const { knex, testCollection, testProvider } = t.context;
+
+  // Create a PDR in the "running" status.
+  const testPdr = generateTestPdr({
+    collectionId: buildCollectionId(testCollection.name, testCollection.version),
+    provider: testProvider.name,
+    status: 'running',
+  });
+
+  await migratePdrRecord(testPdr, knex);
+
+  // An upsert of a PDR in the "running" state may return 0 rows if the upsert
+  // conditions are not met. In this case the migration of this record will fail.
+  const newerTestPdr = {
+    ...testPdr,
+    updatedAt: Date.now(),
+  };
+
+  await t.throwsAsync(
+    migratePdrRecord(newerTestPdr, knex),
+    { instanceOf: PostgresUpdateFailed }
+  );
+});
+
+test.serial('migratePdrRecord updates an already migrated record if the updated date is newer', async (t) => {
+  const {
+    knex,
+    testCollection,
+    testProvider,
+    pdrPgModel,
+  } = t.context;
+
+  const executionPgModel = new ExecutionPgModel();
+  const execution = fakeExecutionRecordFactory();
+
+  await executionPgModel.create(
+    knex,
+    execution
+  );
+
+  const testPdr = generateTestPdr({
+    collectionId: buildCollectionId(testCollection.name, testCollection.version),
+    provider: testProvider.name,
+    execution: execution.arn,
+    status: 'completed',
+    updatedAt: Date.now() - 1000,
+  });
+  await migratePdrRecord(testPdr, knex);
+
+  const newerTestPdr = {
+    ...testPdr,
+    updatedAt: Date.now(),
+  };
+
+  await migratePdrRecord(newerTestPdr, knex);
+
+  const createdRecord = await pdrPgModel.get(knex, { name: testPdr.pdrName });
+
+  t.deepEqual(createdRecord.updated_at, new Date(newerTestPdr.updatedAt));
 });
 
 test.serial('migratePdrs skips already migrated record', async (t) => {
@@ -243,7 +323,10 @@ test.serial('migratePdrs skips already migrated record', async (t) => {
     testCollection,
     testProvider,
   } = t.context;
-  const testPdr = generateTestPdr(testCollection, testProvider.name);
+  const testPdr = generateTestPdr({
+    collectionId: buildCollectionId(testCollection.name, testCollection.version),
+    provider: testProvider.name,
+  });
 
   await migratePdrRecord(testPdr, knex);
   await pdrsModel.create(testPdr);
@@ -269,8 +352,14 @@ test.serial('migratePdrs processes multiple PDR records', async (t) => {
     testProvider,
   } = t.context;
 
-  const testPdr = generateTestPdr(testCollection, testProvider.name);
-  const anotherPdr = generateTestPdr(testCollection, testProvider.name);
+  const testPdr = generateTestPdr({
+    collectionId: buildCollectionId(testCollection.name, testCollection.version),
+    provider: testProvider.name,
+  });
+  const anotherPdr = generateTestPdr({
+    collectionId: buildCollectionId(testCollection.name, testCollection.version),
+    provider: testProvider.name,
+  });
 
   await Promise.all([
     pdrsModel.create(testPdr),
@@ -298,8 +387,14 @@ test.serial('migratePdrs processes all non-failing records', async (t) => {
     testProvider,
   } = t.context;
 
-  const testPdr = generateTestPdr(testCollection, testProvider.name);
-  const anotherPdr = generateTestPdr(testCollection, testProvider.name);
+  const testPdr = generateTestPdr({
+    collectionId: buildCollectionId(testCollection.name, testCollection.version),
+    provider: testProvider.name,
+  });
+  const anotherPdr = generateTestPdr({
+    collectionId: buildCollectionId(testCollection.name, testCollection.version),
+    provider: testProvider.name,
+  });
   delete testPdr.status;
 
   await Promise.all([